fix: 홈페이지 워크스페이스 수정 시 유저 정보 오류 수정
