+++ src/app/(providers)/(root)/[workspaceId]/page.tsx
@@ -1,8 +1,11 @@
 'use client';
+import api from '@/api';
 import useWorkspaceId from '@/hooks/useWorkspaceId';
 import useWorkspaceUser from '@/hooks/useWorkspaceUser';
 import useWorkspaceUserList from '@/hooks/useWorkspaceUserList';
 import useUserStore from '@/store/userStore';
+import { TWorkspaceInfo } from '@/types/workspace';
+import { useEffect, useState } from 'react';
 import { useShallow } from 'zustand/react/shallow';
 import Header from './_components/Header';
 import HomeMemberCard from './_components/HomeMemberCard';
@@ -11,14 +14,28 @@ import MemberNotExistComponent from './_components/MemberNotExistComponent';
 
 const Homepage = () => {
   const workspaceId = useWorkspaceId();
-  const { workspaceUserId, workspaceList } = useUserStore(
+  const [workspaceUserId, setWorkspaceUserId] = useState('');
+  const [workspaceList, setWorkspaceList] = useState<TWorkspaceInfo[] | undefined>();
+  const { userId, setWorkspaceData } = useUserStore(
     useShallow((state) => ({
-      workspaceUserId: state.workspaceUserId,
-      workspaceList: state.workspaceList
+      userId: state.userId,
+      setWorkspaceData: state.setWorkspaceData
     }))
   );
   const { workspaceUser } = useWorkspaceUser(workspaceUserId);
   const { workspaceUserList } = useWorkspaceUserList(workspaceId);
+
+  const getWorkspacaeList = async (workspaceId: number, userId: string) => {
+    const data = await api.workspaceList.getWorkspaceList(workspaceId, userId);
+    const workspaceUserId = data.userData[0].id;
+    const workspaceList = data.workspaceListData;
+    setWorkspaceData(workspaceUserId, workspaceList);
+    setWorkspaceUserId(workspaceUserId);
+    setWorkspaceList(workspaceList);
+  };
+  useEffect(() => {
+    if (workspaceId && userId) getWorkspacaeList(workspaceId, userId);
+  }, [userId, workspaceId]);
 
   if (!(workspaceUser && workspaceList && workspaceUserList)) return;
 
@@ -26,7 +43,11 @@ const Homepage = () => {
     <div>
       <Header workspaceList={workspaceList} workspaceId={workspaceId} />
       <main className="px-[16px] mt-[26px]">
-        <HomeMemberCard name={workspaceUser.name} status={workspaceUser.state} />
+        <HomeMemberCard
+          profileImg={workspaceUser.profile_image}
+          name={workspaceUser.name}
+          status={workspaceUser.state}
+        />
 
         {workspaceUserList.length === 0 ? (
           <MemberNotExistComponent />
